Make this query use an index, because bringing the site to its knees for a minute or two isn't all it's cracked up to be.
git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@4046 1dfba125-df10-0410-8a77-af1a8d6d074f

+++ cgi-bin/ljlib.pl
@@ -4149,7 +4149,7 @@ sub expunge_userpic {
     return undef unless $dbh;
     $dbh->do("UPDATE userpic SET state='X' WHERE picid = ?", undef, $picid);
     return LJ::error($dbh) if $dbh->err;
-    $dbh->do("DELETE FROM userpicmap WHERE picid = ?", undef, $picid);
+    $dbh->do("DELETE FROM userpicmap WHERE userid = ? AND picid = ?", undef, $userid, $picid);
 
     # now clear the user's memcache picture info
     LJ::MemCache::delete([$userid, "upicinf:$userid"]);
